pyproject.toml
[tool.poetry]
name = "NiimPrintX"
version = "0.4.0"
description = "App to print labels using NiimBot printer"
authors = ["Labbots <labbots@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
packages = [{include = "NiimPrintX"}]

[tool.poetry.scripts]
niimprintx = "NiimPrintX.cli.command:niimbot_cli"
niimprintx-ui = "NiimPrintX.ui.__main__:main"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
click = "^8.1.7"
# TODO: bleak 0.22.x is end-of-life; plan migration to bleak >=1.0 (API breaking changes in connect/disconnect return types)
bleak = "^0.22.3"
loguru = "^0.7.2"
pillow = "^12.1.1"
platformdirs = "^4.0"
rich = "^13.7.1"
pycairo = {version = "^1.26.0", optional = true}
wand = {version = "^0.6.13", optional = true}
sv-ttk = {version = "^2.6", optional = true}

[tool.poetry.extras]
gui = ["sv-ttk", "pycairo", "wand"]

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.19"
pytest = "^8.0"
pytest-asyncio = ">=0.23,<2"
ruff = "^0.15"
pytest-cov = "^5.0"
mypy = "^1.11"
types-Pillow = "^10.2"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.report]
fail_under = 60

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
